+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/shared/features/HubdleGradleFeatureExtension.kt
@@ -1,29 +1,36 @@
 package com.javiersc.hubdle.project.extensions.shared.features
 
 import com.javiersc.hubdle.project.extensions.HubdleDslMarker
+import com.javiersc.hubdle.project.extensions._internal.Configurable.Priority
 import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
 import com.javiersc.hubdle.project.extensions.java.hubdleJava
 import com.javiersc.hubdle.project.extensions.kotlin.jvm.hubdleKotlinJvm
 import com.javiersc.hubdle.project.extensions.shared.features.gradle.HubdleGradlePluginFeatureExtension
 import com.javiersc.hubdle.project.extensions.shared.features.gradle.HubdleGradleVersionCatalogFeatureExtension
+import java.io.File
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
+import org.gradle.api.initialization.dsl.VersionCatalogBuilder
+import org.gradle.api.plugins.catalog.CatalogPluginExtension
 import org.gradle.api.provider.Property
 
 public open class HubdleGradleFeatureExtension
 @Inject
 constructor(
     project: Project,
-) : HubdleEnableableExtension(project) {
+) : HubdleConfigurableExtension(project) {
 
     override val isEnabled: Property<Boolean> = property { false }
 
     override val oneOfExtensions: Set<HubdleEnableableExtension>
         get() = setOf(hubdleJava, hubdleKotlinJvm)
+
+    override val priority: Priority = Priority.P3
 
     public val plugin: HubdleGradlePluginFeatureExtension
         get() = getHubdleExtension()
@@ -33,15 +40,26 @@ constructor(
         plugin.enableAndExecute(action)
     }
 
-    public val versionCatalog: HubdleGradleVersionCatalogFeatureExtension
+    private val versionCatalog: HubdleGradleVersionCatalogFeatureExtension
         get() = getHubdleExtension()
 
     @HubdleDslMarker
-    public fun versionCatalog(
-        action: Action<HubdleGradleVersionCatalogFeatureExtension> = Action {}
-    ) {
-        versionCatalog.enableAndExecute(action)
+    public fun versionCatalogs(action: Action<CatalogPluginExtension> = Action {}) {
+        versionCatalog.enabled(true)
+        configurable { action.execute(the()) }
     }
+
+    @HubdleDslMarker
+    public fun CatalogPluginExtension.catalog(action: Action<VersionCatalogBuilder> = Action {}) {
+        this.versionCatalog(action::execute)
+    }
+
+    @HubdleDslMarker
+    public fun VersionCatalogBuilder.toml(toml: File) {
+        from(project.files(toml))
+    }
+
+    override fun Project.defaultConfiguration() {}
 }
 
 public interface HubdleGradleDelegateFeatureExtension : BaseHubdleDelegateExtension {