+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/shared/features/gradle/HubdleGradleVersionCatalogFeatureExtension.kt
@@ -1,6 +1,5 @@
 package com.javiersc.hubdle.project.extensions.shared.features.gradle
 
-import com.javiersc.hubdle.project.extensions.HubdleDslMarker
 import com.javiersc.hubdle.project.extensions._internal.ApplicablePlugin.Scope
 import com.javiersc.hubdle.project.extensions._internal.Configurable.Priority
 import com.javiersc.hubdle.project.extensions._internal.PluginId
@@ -10,13 +9,9 @@ import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.config.publishing.maven.configurableMavenPublishing
 import com.javiersc.hubdle.project.extensions.java.hubdleJava
 import com.javiersc.hubdle.project.extensions.kotlin.jvm.hubdleKotlinJvm
-import java.io.File
 import javax.inject.Inject
 import org.gradle.api.Project
-import org.gradle.api.file.ConfigurableFileCollection
-import org.gradle.api.plugins.catalog.CatalogPluginExtension
 import org.gradle.api.provider.Property
-import org.gradle.kotlin.dsl.configure
 
 public open class HubdleGradleVersionCatalogFeatureExtension
 @Inject
@@ -31,25 +26,12 @@ constructor(
 
     override val priority: Priority = Priority.P3
 
-    public val catalogs: ConfigurableFileCollection = project.files(emptyList<File>())
-
-    @HubdleDslMarker
-    public fun catalogs(vararg files: File) {
-        catalogs.setFrom(files)
-    }
-
     override fun Project.defaultConfiguration() {
         applicablePlugin(
             priority = Priority.P3,
             scope = Scope.CurrentProject,
             PluginId.GradleVersionCatalog,
         )
-
-        configurable {
-            configure<CatalogPluginExtension> {
-                versionCatalog { catalog -> catalog.from(catalogs) }
-            }
-        }
 
         configurableMavenPublishing(mavenPublicationName = "versionCatalog")
     }